+++ gegelatilib/include/gegelati.h
@@ -65,6 +65,10 @@
 #include <learn/learningParameters.h>
 #include <learn/parallelLearningAgent.h>
 
+#include <learn/adversarialEvaluationResult.h>
+#include <learn/adversarialLearningAgent.h>
+#include <learn/adversarialLearningEnvironment.h>
+
 #include <learn/classificationEvaluationResult.h>
 #include <learn/classificationLearningAgent.h>
 #include <learn/classificationLearningEnvironment.h>
@@ -73,6 +77,8 @@
 #include <log/laLogger.h>
 #include <log/laPolicyStatsLogger.h>
 #include <log/logger.h>
+
+#include <learn/job.h>
 
 #include <mutator/lineMutator.h>
 #include <mutator/mutationParameters.h>