+++ test/classificationLearningAgentTest.cpp
@@ -125,8 +125,8 @@ TEST_F(ClassificationLearningAgentTest, EvaluateRoot)
 
     cla.init();
     std::shared_ptr<Learn::EvaluationResult> result1;
-    ASSERT_NO_THROW(result1 = cla.evaluateRoot(
-                        tee, *cla.getTPGGraph().getRootVertices().at(0), 0,
+    ASSERT_NO_THROW(result1 = cla.evaluateJob(
+                        tee, *cla.makeJob(0, Learn::LearningMode::TRAINING), 0,
                         Learn::LearningMode::TRAINING, fle))
         << "Evaluation from a root failed.";
     ASSERT_LE(result1->getResult(), 1.0)
@@ -138,8 +138,8 @@ TEST_F(ClassificationLearningAgentTest, EvaluateRoot)
 
     // Reevaluate to check that the previous result1 is not returned.
     std::shared_ptr<Learn::EvaluationResult> result2;
-    ASSERT_NO_THROW(result2 = cla.evaluateRoot(
-                        tee, *cla.getTPGGraph().getRootVertices().at(0), 0,
+    ASSERT_NO_THROW(result2 = cla.evaluateJob(
+                        tee, *cla.makeJob(0, Learn::LearningMode::TRAINING), 0,
                         Learn::LearningMode::TRAINING, fle))
         << "Evaluation from a root failed.";
     ASSERT_NE(result1, result2);
@@ -150,8 +150,8 @@ TEST_F(ClassificationLearningAgentTest, EvaluateRoot)
 
     // Reevaluate to check that the previous result2 is returned.
     std::shared_ptr<Learn::EvaluationResult> result3;
-    ASSERT_NO_THROW(result3 = cla.evaluateRoot(
-                        tee, *cla.getTPGGraph().getRootVertices().at(0), 0,
+    ASSERT_NO_THROW(result3 = cla.evaluateJob(
+                        tee, *cla.makeJob(0, Learn::LearningMode::TRAINING), 0,
                         Learn::LearningMode::TRAINING, fle))
         << "Evaluation from a root failed.";
     ASSERT_EQ(result3, result2);